@@ -4,7 +4,7 @@ package app.septs.idrw
 import android.content.Context
 import android.content.Intent
 import android.content.pm.ActivityInfo
-import android.databinding.DataBindingUtil
+import androidx.databinding.DataBindingUtil
 import android.net.Uri
 import android.os.Bundle
 import android.os.Looper
